+++ NEWS.md
@@ -12,6 +12,7 @@
 6. Fix #33 - Confirm logic in `current_hosp_data()` (@rjake)
 7. Fix #38 - Add function `current_hai_data()`
 8, Fix #39 - Add function `current_hcahps_data()`
+9. Fix #42 - Add function `current_hvbp_data()`
 
 ## Minor Fixes and Improvements
 None